+++ source/game_sa/Conversations/ConversationForPed.cpp
@@ -1,0 +1,30 @@
+#include "ConversationForPed.h"
+#include "Conversations.h"
+
+void CConversationForPed::InjectHooks() {
+    RH_ScopedClass(CConversationForPed);
+    RH_ScopedCategory("Conversations");
+    RH_ScopedInstall(Update, 0x43C190, {.reversed = false});
+    RH_ScopedInstall(IsPlayerInPositionForConversation, 0x43AC40);
+}
+
+inline void CConversationForPed::Clear(bool dontClearNodes) {
+    if (!dontClearNodes) {
+        CConversations::m_Nodes[m_FirstNode].ClearRecursively();
+    }
+    m_FirstNode                 = -1;
+    m_CurrentNode               = -1;
+    m_pPed                      = nullptr;
+    m_LastChange                = 0;
+    m_LastTimeWeWereCloseEnough = 0;
+}
+
+// 0x43C190
+void CConversationForPed::Update() {
+    plugin::CallMethod<0x43C190, CConversationForPed*>(this);
+}
+
+// 0x43AC40
+bool CConversationForPed::IsPlayerInPositionForConversation(bool randomConversation) {
+    return plugin::CallMethodAndReturn<bool, 0x43AC40, CConversationForPed*, bool>(this, randomConversation);
+}
