+++ source/game_sa/Conversations/ConversationNode.h
@@ -1,0 +1,17 @@
+#pragma once
+
+class CConversationNode {
+public:
+    char  m_Name[8];
+    int16 m_NodeYes;
+    int16 m_NodeNo;
+    int32 m_Speech;
+    int32 m_SpeechY;
+    int32 m_SpeechN;
+
+    static void InjectHooks();
+    inline void Clear();
+    inline void ClearRecursively();
+};
+
+VALIDATE_SIZE(CConversationNode, 0x18);
