+++ source/game_sa/Conversations/Conversations.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#include "ConversationForPed.h"
+#include "PedToPlayerConversations.h"
+#include "ConversationNode.h"
+
 class CPed;
 
 enum {
@@ -7,46 +11,6 @@ enum {
     MAX_NUM_CONVERSATION_NODES = 50,
     MAX_NUM_TEMP_CONVERSATION_NODES = 12,
 };
-
-
-class CPedToPlayerConversations {
-public:
-    enum class eP2pState : uint32 {
-        INACTIVE = 0,
-        PEDHASOPENED,
-        WAITINGFORFINALWORD,
-        WAITINGTOFINISH,
-    };
-
-    static inline eP2pState& m_State                        = StaticRef<eP2pState, 0x969A20>();
-    static inline CPed*&     m_pPed                         = StaticRef<CPed*, 0x9691C0>();
-    static inline int32&     m_Topic                        = StaticRef<int32, 0x9691BC>();
-    static inline uint32&    m_TimeOfLastPlayerConversation = StaticRef<uint32, 0x9691B4>();
-    static inline uint32&    m_StartTime                    = StaticRef<uint32, 0x9691B8>();
-    static inline bool&      m_bPositiveReply               = StaticRef<bool, 0x9691B0>(); // unused
-    static inline bool&      m_bPositiveOpening             = StaticRef<bool, 0x9691B1>();
-
-    static void InjectHooks();
-    static void Clear();
-    static void Update();
-    static void EndConversation();
-};
-
-class CConversationNode {
-public:
-    char  m_Name[8];
-    int16 m_NodeYes;
-    int16 m_NodeNo;
-    int32 m_Speech;
-    int32 m_SpeechY;
-    int32 m_SpeechN;
-
-    static void InjectHooks();
-    inline void Clear();
-    inline void ClearRecursively();
-};
-
-VALIDATE_SIZE(CConversationNode, 0x18);
 
 class CTempConversationNode {
 public:
@@ -63,32 +27,6 @@ public:
 };
 
 VALIDATE_SIZE(CTempConversationNode, 0x2C);
-
-class CConversationForPed {
-public:
-    enum eStatus : uint32 {
-        INACTIVE = 0,
-        PLAYER_SPEAKING,
-        PED_SPEAKING,
-        WAITINGFORINPUT
-    };
-
-    int32   m_FirstNode;
-    int32   m_CurrentNode;
-    CPed*   m_pPed;
-    uint32  m_LastChange;
-    uint32  m_LastTimeWeWereCloseEnough;
-    eStatus m_Status;
-    bool    m_Enabled;
-    bool    m_SuppressSubtitles;
-
-    static void InjectHooks();
-    inline void Clear(bool dontClearNodes);
-    void        Update();
-    bool        IsPlayerInPositionForConversation(bool randomConversation);
-};
-
-VALIDATE_SIZE(CConversationForPed, 0x1C);
 
 class CConversations {
 public:
