+++ source/game_sa/Conversations/PedToPlayerConversations.cpp
@@ -1,0 +1,48 @@
+#include "PedToPlayerConversations.h"
+#include "IKChainManager_c.h"
+
+void CPedToPlayerConversations::InjectHooks() {
+    RH_ScopedClass(CPedToPlayerConversations);
+    RH_ScopedCategory("Conversations");
+    RH_ScopedInstall(Clear, 0x43AAE0);
+    RH_ScopedInstall(Update, 0x43B0F0, {.reversed = false});
+    RH_ScopedInstall(EndConversation, 0x43AB10);
+}
+
+// 0x43AAE0
+void CPedToPlayerConversations::Clear() {
+    ZoneScoped;
+
+    if (m_State != eP2pState::INACTIVE) {
+        CAEPedSpeechAudioEntity::ReleasePlayerConversation();
+        m_State = eP2pState::INACTIVE;
+    }
+
+    m_pPed                         = nullptr;
+    m_TimeOfLastPlayerConversation = 0;
+}
+
+// 0x43B0F0
+void CPedToPlayerConversations::Update() {
+    plugin::Call<0x43B0F0>();
+}
+
+// 0x43AB10
+void CPedToPlayerConversations::EndConversation() {
+    m_State           = eP2pState::INACTIVE;
+
+    CAEPedSpeechAudioEntity::ReleasePlayerConversation();
+
+    if (m_pPed) {
+        m_pPed->EnablePedSpeech();
+    }
+
+    const auto player = FindPlayerPed(-1);
+    if (g_ikChainMan.IsLooking(player)) {
+        g_ikChainMan.AbortLookAt(player, 250);
+    }
+
+    if (m_pPed && g_ikChainMan.IsLooking(m_pPed)) {
+        return g_ikChainMan.AbortLookAt(m_pPed, 250);
+    }
+}
